cleanup & inline suppressible plugin issue
