--- src/Handler/ClassHandler.php
@@ -1,73 +1,0 @@
-<?php
-
-namespace Psalm\SymfonyPsalmPlugin\Handler;
-
-use PhpParser\Node;
-use PhpParser\Node\Expr;
-use Psalm\Codebase;
-use Psalm\CodeLocation;
-use Psalm\Context;
-use Psalm\IssueBuffer;
-use Psalm\Plugin\Hook\AfterClassLikeAnalysisInterface;
-use Psalm\Plugin\Hook\AfterMethodCallAnalysisInterface;
-use Psalm\StatementsSource;
-use Psalm\Storage\ClassLikeStorage;
-use Psalm\SymfonyPsalmPlugin\Issue\ContainerDependency;
-use Psalm\Type\Atomic\TArray;
-use Psalm\Type\Atomic\TInt;
-use Psalm\Type\Atomic\TNull;
-use Psalm\Type\Atomic\TString;
-use Psalm\Type\Union;
-use Symfony\Component\DependencyInjection\ContainerInterface;
-
-class ClassHandler implements AfterClassLikeAnalysisInterface, AfterMethodCallAnalysisInterface
-{
-    /**
-     * {@inheritdoc}
-     */
-    public static function afterStatementAnalysis(Node\Stmt\ClassLike $stmt, ClassLikeStorage $classlike_storage, StatementsSource $statements_source, Codebase $codebase, array &$file_replacements = [])
-    {
-        foreach ($stmt->stmts as $stmt) {
-            if ($stmt instanceof Node\Stmt\ClassMethod && '__construct' === $stmt->name->name) {
-                foreach ($stmt->params as $param) {
-                    if ($param->type instanceof Node\Name && ContainerInterface::class === $param->type->getAttributes()['resolvedName']) {
-                        IssueBuffer::accepts(
-                            new ContainerDependency(new CodeLocation($statements_source, $stmt)),
-                            $statements_source->getSuppressedIssues()
-                        );
-                    }
-                }
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function afterMethodCallAnalysis(
-        Expr $expr,
-        string $method_id,
-        string $appearing_method_id,
-        string $declaring_method_id,
-        Context $context,
-        StatementsSource $statements_source,
-        Codebase $codebase,
-        array &$file_replacements = [],
-        Union &$return_type_candidate = null
-    ) {
-        switch ($declaring_method_id) {
-            case 'Symfony\Component\HttpFoundation\HeaderBag::get':
-                if ($return_type_candidate) {
-                    /** @psalm-suppress MixedArrayAccess */
-                    if (isset($expr->args[2]->value->name->parts[0]) && 'false' === $expr->args[2]->value->name->parts[0]) {
-                        $return_type_candidate = new Union([new TArray([new Union([new TInt()]), new Union([new TString()])])]);
-                    } else {
-                        $return_type_candidate = new Union([new TString(), new TNull()]);
-                    }
-                }
-                break;
-        }
-    }
-}
