+++ src/Plugin.php
@@ -5,10 +5,11 @@ namespace Psalm\SymfonyPsalmPlugin;
 use Doctrine\Common\Annotations\AnnotationRegistry;
 use Psalm\Plugin\PluginEntryPointInterface;
 use Psalm\Plugin\RegistrationInterface;
-use Psalm\SymfonyPsalmPlugin\Handler\ClassHandler;
 use Psalm\SymfonyPsalmPlugin\Handler\ConsoleHandler;
+use Psalm\SymfonyPsalmPlugin\Handler\ContainerDependencyHandler;
 use Psalm\SymfonyPsalmPlugin\Handler\ContainerHandler;
 use Psalm\SymfonyPsalmPlugin\Handler\DoctrineRepositoryHandler;
+use Psalm\SymfonyPsalmPlugin\Handler\HeaderBagHandler;
 use Psalm\SymfonyPsalmPlugin\Symfony\ContainerMeta;
 use SimpleXMLElement;
 
@@ -19,13 +20,15 @@ class Plugin implements PluginEntryPointInterface
      */
     public function __invoke(RegistrationInterface $api, SimpleXMLElement $config = null)
     {
-        require_once __DIR__.'/Handler/ClassHandler.php';
+        require_once __DIR__.'/Handler/HeaderBagHandler.php';
         require_once __DIR__.'/Handler/ContainerHandler.php';
         require_once __DIR__.'/Handler/ConsoleHandler.php';
         require_once __DIR__.'/Handler/DoctrineRepositoryHandler.php';
+        require_once __DIR__.'/Handler/ContainerDependencyHandler.php';
 
-        $api->registerHooksFromClass(ClassHandler::class);
+        $api->registerHooksFromClass(HeaderBagHandler::class);
         $api->registerHooksFromClass(ConsoleHandler::class);
+        $api->registerHooksFromClass(ContainerDependencyHandler::class);
 
         if (class_exists(AnnotationRegistry::class)) {
             /** @psalm-suppress DeprecatedMethod */
