+++ tests/acceptance/acceptance/ContainerDependency.feature
@@ -35,3 +35,19 @@ Feature: ContainerDependency
       | Type                 | Message                                                   |
       | ContainerDependency | Container must not inject into services as dependency! |
     And I see no other errors
+
+  Scenario: Asserting container dependency issue can be suppressed inline
+    Given I have the following code
+      """
+      <?php
+      use Symfony\Component\DependencyInjection\ContainerInterface;
+      class SomeService
+      {
+        /** @psalm-suppress ContainerDependency */
+        public function __construct(ContainerInterface $container)
+        {
+        }
+      }
+      """
+    When I run Psalm
+    Then I see no errors
